refactor(react): rename webview to webview-app
