+++ packages/botonic-react/src/index.js
@@ -7,4 +7,4 @@ export { NodeApp } from './node-app'
 export { staticAsset } from './util/environment'
 export { getBotonicApp, Webchat } from './webchat/index.js'
 export { WebchatApp } from './webchat-app'
-export { WebviewApp } from './webview'
+export { WebviewApp } from './webview-app'
